babappaomegav1.3.0/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "babappai"
version = "2.0.0"
description = "BABAPPAi: diagnostic framework for identifiability of episodic branch-site structure"
readme = "README.md"
requires-python = ">=3.9"
license = "MIT"
authors = [{ name = "Krishnendu Sinha" }]

dependencies = [
  "torch>=2.0",
  "numpy",
  "biopython",
  "ete3",
  "six",
  "platformdirs"
]

keywords = [
  "episodic selection",
  "branch-site",
  "identifiability",
  "evolutionary genomics",
  "likelihood-free inference",
  "deep learning"
]

classifiers = [
  "Development Status :: 5 - Production/Stable",
  "Intended Audience :: Science/Research",
  "Topic :: Scientific/Engineering :: Bio-Informatics",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Operating System :: OS Independent"
]

[project.optional-dependencies]
test = [
  "pytest>=8",
  "pytest-mock>=3.14"
]
dev = [
  "build>=1.2",
  "twine>=5",
  "pytest>=8"
]

[project.urls]
Homepage = "https://github.com/krishnendusinha/babappai"
Documentation = "https://github.com/krishnendusinha/babappai"
Source = "https://github.com/krishnendusinha/babappai"
LegacySoftwareDOI = "https://doi.org/10.5281/zenodo.18520163"
LegacyModelDOI = "https://doi.org/10.5281/zenodo.18195869"

[project.scripts]
babappai = "babappai.cli:main"
babappaomega = "babappaomega.cli:main"

[tool.setuptools]
packages = [
  "babappai",
  "babappai.assets",
  "babappai.data",
  "babappai.calibration",
  "babappai.validation",
  "babappaomega",
  "babappaomega.assets",
  "babappaomega.data"
]

[tool.setuptools.package-data]
babappai = [
  "py.typed",
  "data/*.json",
  "assets/*.json"
]
babappaomega = [
  "py.typed",
  "data/*.json",
  "assets/*.json"
]

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]
